pyproject.toml
[project]
name = "redis-openai-agents"
version = "0.1.0"
description = "Redis integrations for OpenAI Agents SDK - session management, vector search, caching, and observability"
authors = [
    {name = "Brian Sam-Bodden", email = "bsb@redis.com"}
]
readme = "README.md"
requires-python = ">=3.10,<3.14"
license = {text = "MIT"}
keywords = ["redis", "openai", "agents", "ai", "rag", "vector-search", "caching"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "openai-agents>=0.14.2",
    "redis[hiredis]>=5.2.0",
    "redisvl>=0.17.1",
    "pydantic>=2.0.0",
    "sentence-transformers>=5.1.2",
]

[project.optional-dependencies]
vector = []
cache = []
search = []
all = []
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
    "testcontainers[redis]>=4.14.2",
]
docs = [
    "sphinx>=7.0.0",
    "pydata-sphinx-theme>=0.15.0",
    "sphinx-design>=0.5.0",
    "sphinx-copybutton>=0.5.0",
    "myst-nb>=1.0.0",
    "nbsphinx>=0.9.0",
    "sphinx-favicon>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/redis/redis-openai-agents"
Documentation = "https://redis.github.io/redis-openai-agents"
Repository = "https://github.com/redis/redis-openai-agents"
Issues = "https://github.com/redis/redis-openai-agents/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
    "testcontainers[redis]>=4.14.2",
    "jupyter>=1.0.0",
    "ipykernel>=6.0.0",
    "sphinx>=7.0.0",
    "pydata-sphinx-theme>=0.15.0",
    "sphinx-design>=0.5.0",
    "sphinx-copybutton>=0.5.0",
    "myst-nb>=1.0.0",
    "nbsphinx>=0.9.0",
    "sphinx-favicon>=1.0.0",
]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "N", "B"]
ignore = ["E501"]  # Line too long (handled by formatter)

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_any_generics = false
disallow_subclassing_any = false

[[tool.mypy.overrides]]
module = "mcp.*"
ignore_errors = true
follow_imports = "skip"

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
addopts = "-v --strict-markers"
filterwarnings = [
    "ignore::DeprecationWarning:testcontainers",
]
markers = [
    "asyncio: mark test as async",
    "integration: mark test as integration test",
]

[tool.coverage.run]
source = ["src"]
omit = ["tests/*", "*/test_*.py"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
]
